+++ src/main/java/com/cisowski/schoolmanagement/timetable/scheduleOccurrence/mapper/ScheduleOccurrenceMapper.java
@@ -8,10 +8,15 @@ import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.Named;
 
+import java.util.List;
+import java.util.Set;
+
 @Mapper(config = BaseMapperConfig.class, uses = {ScheduleMapper.class})
 public interface ScheduleOccurrenceMapper {
 
     @Named("toScheduleOccurrenceSummaryResponse")
     @Mapping(target = "schedule", source = "schedule", qualifiedByName = "toScheduleSummaryResponse")
     ScheduleOccurrenceSummaryResponse toSummaryResponse(ScheduleOccurrenceEntity entity);
+
+    List<ScheduleOccurrenceSummaryResponse> toSummaryResponseList(Set<ScheduleOccurrenceEntity> entities);
 }