+++ src/main/java/com/cisowski/schoolmanagement/timetable/scheduleOccurrence/repository/ScheduleOccurrenceRepository.java
@@ -11,6 +11,7 @@ import org.springframework.stereotype.Repository;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 
 @Repository
 public interface ScheduleOccurrenceRepository extends JpaRepository<ScheduleOccurrenceEntity, Long> {
@@ -25,4 +26,6 @@ public interface ScheduleOccurrenceRepository extends JpaRepository<ScheduleOccu
             @Param("timeNow") LocalDateTime timeNow,
             @Param("timeThreshold") LocalDateTime timeThreshold
     );
+
+    Set<ScheduleOccurrenceEntity> findByScheduleInAndOccurrenceDateTimeBetween(List<ScheduleEntity> schedules, LocalDateTime thresholdStartTime, LocalDateTime thresholdEndTime);
 }