+++ src/main/java/com/cisowski/schoolmanagement/timetable/scheduleOccurrence/service/ScheduleOccurrenceService.java
@@ -3,6 +3,7 @@ package com.cisowski.schoolmanagement.timetable.scheduleOccurrence.service;
 import com.cisowski.schoolmanagement.timetable.schedule.model.ScheduleEntity;
 import com.cisowski.schoolmanagement.timetable.scheduleOccurrence.model.OccurrenceStatus;
 import com.cisowski.schoolmanagement.timetable.scheduleOccurrence.model.ScheduleOccurrenceEntity;
+import com.cisowski.schoolmanagement.timetable.scheduleOccurrence.model.ScheduleOccurrenceSummaryResponse;
 
 import java.util.List;
 
@@ -11,4 +12,7 @@ public interface ScheduleOccurrenceService {
     void initializeScheduleOccurrence(ScheduleEntity schedule, Integer minInitDays);
     List<ScheduleOccurrenceEntity> fetchUninitializedOccurrencesForAttendance();
     void changeOccurrenceStatus(ScheduleOccurrenceEntity occurrence, OccurrenceStatus occurrenceStatus);
+    List<ScheduleOccurrenceSummaryResponse> getOccurrencesForYearbook(Integer yearbookId);
+    ScheduleOccurrenceSummaryResponse getOccurrenceById(Long scheduleOccurrenceId);
+    ScheduleOccurrenceEntity fetchOccurrence(Long scheduleOccurrenceId);
 }